+++ shared/core/data/src/commonMain/kotlin/com/yellowtrack/platform/core/data/DataModule.kt
@@ -10,6 +10,7 @@ import com.yellowtrack.platform.core.data.internal.SqlDelightDeliverableReposito
 import com.yellowtrack.platform.core.data.internal.SqlDelightExpenseRepository
 import com.yellowtrack.platform.core.data.internal.SqlDelightInvoiceRepository
 import com.yellowtrack.platform.core.data.internal.SqlDelightLeadRepository
+import com.yellowtrack.platform.core.data.internal.SqlDelightMediaCopyRepository
 import com.yellowtrack.platform.core.data.internal.SqlDelightPostProductionRepository
 import com.yellowtrack.platform.core.data.internal.SqlDelightProjectRepository
 import com.yellowtrack.platform.core.data.internal.SqlDelightQuoteRepository
@@ -42,6 +43,7 @@ val dataModule =
         single<TalentReleaseRepository> { SqlDelightTalentReleaseRepository(get(), get(), ioDispatcher) }
         single<PostProductionRepository> { SqlDelightPostProductionRepository(get(), get(), get(), ioDispatcher) }
         single<DeliverableRepository> { SqlDelightDeliverableRepository(get(), get(), ioDispatcher) }
+        single<MediaCopyRepository> { SqlDelightMediaCopyRepository(get(), get(), ioDispatcher) }
         single<ServiceTemplateRepository> { SqlDelightServiceTemplateRepository(get(), get(), get(), ioDispatcher) }
 
         single<LeadRepository> { SqlDelightLeadRepository(get(), get(), get(), ioDispatcher) }